backend-legacy/requirements.txt
# Core Flask dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.0
Flask-SocketIO>=5.3.0

# Database
pymongo>=4.5.0

# Environment and configuration
python-dotenv>=1.0.0

# AI and OpenAI
openai>=1.3.0

# PDF generation
reportlab>=4.0.0

# Payment processing
stripe>=7.0.0

# File handling
Werkzeug>=2.3.0

# Audio processing (Python 3.12 compatible versions)
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.26.0
pydub>=0.25.0

# Additional audio dependencies
ffmpeg-python>=0.2.0